requirements.txt
streamlit==1.39.0
transformers==4.45.1
sentence-transformers==2.2.2
torch>=2.0.0
nltk==3.9.1
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==0.6.23
scikit-learn==1.5.2
# sumy==0.11.0
yake==0.4.8
networkx
matplotlib
pyvis

